feat: Add the ability to play with Smith notation
This iteration further refine the core game loop. It is now possible to
play with Smith notation in two input variations.
Preview mode: Enter the piece location only will trigger the mode
indicator. A second prompt will follow to collect the move position.
In honour to classic chess rule, it is by design not possible to preview
another piece.

Select and move: Enter the full valid sequence will select and move the
the piece with one prompt.